requirements.txt
# TradingView entegrasyonu
tradingview-ta==3.3.0
python-dotenv==1.0.0

# Veri işleme
pandas==2.1.1
numpy==1.26.0

# API ve veri çekme
requests==2.31.0
aiohttp==3.8.6
ccxt==4.1.0  # Kripto exchange entegrasyonu

# Teknik analiz
ta==0.11.0
pandas-ta==0.3.14b0

# Grafik ve görselleştirme
matplotlib==3.8.0
plotly==5.17.0

# Loglama
loguru==0.7.2

# Test
pytest==7.4.2
pytest-asyncio==0.21.1
